initState() and super.initState() Section 2, Lecture 33
In the last lecture, initState() was introduced. You learned that you have to call super.initState() in there and that you should put it at the end of your initState() implementation.

It's now actually recommend to move at at the beginning, though it will not influence the way your code runs. So the code shown in this course will still work fine without any issues.

But to follow the official recommendation and code style, you should use this code:

    @override
      void initState() {
        super.initState();
        _products.add(widget.startingProduct);
      }